# src/ControlPlane/Bagatka.AiSloth.WebApi/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "Bagatka.AiSloth.WebApi | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/workspaces/{workspaceId}/providers": {
      "get": {
        "tags": [
          "Nooks"
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProviderSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/workspaces/{workspaceId}/nooks": {
      "post": {
        "tags": [
          "Nooks"
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NookSummary"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Nooks"
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfNookSummary"
                }
              }
            }
          }
        }
      }
    },
    "/nooks/{id}": {
      "get": {
        "tags": [
          "Nooks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NookSummary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Nooks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/nooks/{id}/download": {
      "get": {
        "tags": [
          "Nooks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "source",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "checkpoint",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/nooks/{nookId}/checkpoints": {
      "get": {
        "tags": [
          "Nooks"
        ],
        "parameters": [
          {
            "name": "nookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfCheckpointSummary"
                }
              }
            }
          }
        }
      }
    },
    "/nooks/{nookId}/setup": {
      "get": {
        "tags": [
          "Nooks"
        ],
        "parameters": [
          {
            "name": "nookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NookSetup"
                }
              }
            }
          }
        }
      }
    },
    "/nooks/{id}/wake": {
      "post": {
        "tags": [
          "Nooks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/nooks/{nookId}/processes": {
      "post": {
        "tags": [
          "Nooks"
        ],
        "parameters": [
          {
            "name": "nookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartProcessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessSummary"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Nooks"
        ],
        "parameters": [
          {
            "name": "nookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfProcessSummary"
                }
              }
            }
          }
        }
      }
    },
    "/nooks/{nookId}/processes/{processId}/output": {
      "get": {
        "tags": [
          "Nooks"
        ],
        "parameters": [
          {
            "name": "nookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "processId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "fromOffset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "required": [
                    "data"
                  ],
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ProcessEvent"
                    },
                    "event": {
                      "enum": [
                        "ProcessOutput",
                        "ProcessExited"
                      ],
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/nooks/{nookId}/processes/{processId}/input": {
      "post": {
        "tags": [
          "Nooks"
        ],
        "parameters": [
          {
            "name": "nookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "processId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendInputRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/nooks/{nookId}/processes/{processId}/stop": {
      "post": {
        "tags": [
          "Nooks"
        ],
        "parameters": [
          {
            "name": "nookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "processId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateNookRequest": {
        "required": [
          "provider"
        ],
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "repositories": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/NookRepository"
            }
          },
          "copyOf": {},
          "checkpoint": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          }
        }
      },
...
      "NookId": {
        "description": "Identifies a nook."
      },
...
      "NookSetup": {
        "required": [
          "scripts",
          "run",
          "readyCopyMadeAt"
        ],
        "type": "object",
        "properties": {
          "scripts": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The scripts, as paths in the nook\u0027s files, in the order they run; empty when there are none."
          },
          "run": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "description": "Their latest run, or \u0060null\u0060 when there are no scripts.",
                "$ref": "#/components/schemas/SetupRun"
              }
            ]
          },
          "readyCopyMadeAt": {
            "type": [
              "null",
              "string"
            ],
            "description": "When the ready copy the latest run set the nook up from was made: a copy of a nook with the same\nfiles right after a setup that took a while, which leaves this setup little to do. \u0060null\u0060\nwhen the run set it up from scratch, or only resumed it.",
            "format": "date-time"
          }
        },
        "description": "A nook\u0027s setup: scripts that come with its files and prepare it, which run whenever the nook gets\nits files, as a new nook, a copy, or one brought back after its sandbox was lost. Every\n\u0060.agents/setup\u0060 runs first, installing what the code needs, then every \u0060.agents/resume\u0060,\nstarting its services; those at the top of the files before each folder\u0027s directly in it, by\nname, each in its own folder, with the workspace\u0027s secrets. A setup may take 30 minutes and\na resume 5. Their output is also written to \u0060/var/log/aisloth/setup.log\u0060 in the nook."
      },
...
      "NookSummary": {
        "required": [
          "id",
          "workspaceId",
          "provider",
          "status",
          "createdAt",
          "usage",
          "diskNearlyFull",
          "image",
          "reservedFor",
          "sources"
        ],
        "type": "object",
        "properties": {
          "id": {
            "description": "The nook.",
            "$ref": "#/components/schemas/NookId"
          },
          "workspaceId": {
            "description": "The workspace that owns it.",
            "$ref": "#/components/schemas/WorkspaceId"
          },
          "provider": {
            "type": "string",
            "description": "The ID of the provider it runs on, as in string ProviderSummary.Id."
          },
          "status": {
            "description": "Where it is in its lifecycle.",
            "$ref": "#/components/schemas/NookStatus"
          },
          "createdAt": {
            "type": "string",
            "description": "When it was recorded.",
            "format": "date-time"
          },
          "usage": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "description": "What it uses of its disk, memory, and CPU while it runs; \u0060null\u0060 while it doesn\u0027t.",
                "$ref": "#/components/schemas/NookUsage"
              }
            ]
          },
          "diskNearlyFull": {
            "type": "boolean",
            "description": "Whether its disk is so full that its work, and keeping its files as checkpoints, may soon fail:\npeople should free space, or move to a nook with a bigger disk."
          },
          "image": {
            "type": [
              "null",
              "string"
            ],
            "description": "The image it started from, such as one with a chat\u0027s harness, or \u0060null\u0060 for the base image."
          },
          "reservedFor": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "description": "The one person who may change what runs in it or its files, such as the owner of the personal\naccount its chat\u0027s agent works on; \u0060null\u0060 for everyone with Write on its workspace.",
                "$ref": "#/components/schemas/UserId"
              }
            ]
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NookSource"
            },
            "description": "Its sources, each a repository at \u0060/work/\u0026lt;name\u0026gt;\u0060, by name."
          }
        },
        "description": "A nook as the people with access to it see it."
      },
...
      "PageOfCheckpointSummary": {
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CheckpointSummary"
            },
            "description": "The items, in the list\u0027s stable order."
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ],
            "description": "Pass it in the next PageRequest; \u0060null\u0060 on the last page."
          }
        },
        "description": "One page of a keyset-paginated list."
      },
      "PageOfNookSummary": {
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NookSummary"
            },
            "description": "The items, in the list\u0027s stable order."
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ],
            "description": "Pass it in the next PageRequest; \u0060null\u0060 on the last page."
          }
        },
        "description": "One page of a keyset-paginated list."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ProcessEvent": {
        "type": "object",
        "anyOf": [
          {
            "$ref": "#/components/schemas/ProcessOutput"
          },
          {
            "$ref": "#/components/schemas/ProcessExited"
          }
        ],
        "description": "Something a running process did: wrote output, or exited. A process\u0027s events arrive in order,\nand ProcessExited is always its last."
      },
...
      "ProviderSummary": {
        "required": [
          "id",
          "name",
          "available"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "What string CreateNook.Provider takes; callers don\u0027t parse it."
          },
          "name": {
            "type": "string",
            "description": "Its display name: \u0060cloud\u0060 for the deployment\u0027s own provider, whatever runs it, or each one\u0027s\nown name, such as \u0060docker\u0060, when a deployment runs several, as in development; a machine\u0027s name\nfor a machine."
          },
          "available": {
            "type": "boolean",
            "description": "Whether it can create nooks right now; a machine that is offline can\u0027t."
          }
        },
        "description": "A provider a workspace\u0027s nooks can run on: one the deployment runs for every workspace, such as a\nDocker Engine or a cloud, or one of the workspace\u0027s own machines."
      },
...
      "StartProcessRequest": {
        "required": [
          "command"
        ],
        "type": "object",
        "properties": {
          "command": {
            "type": "string"
          },
          "arguments": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "workingDirectory": {
            "type": [
              "null",
              "string"
            ]
          },
          "retention": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OutputRetention"
              }
            ]
          },
          "environment": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "WorkspaceId": {},
...
    }
  },
...
}